requirements.txt
# Data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
lxml>=4.9.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
statsmodels>=0.14.0

# ML
scikit-learn>=1.3.0
shap>=0.43.0
optuna>=3.4.0

# Visualization
plotly>=5.18.0

# App
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0